heart_disease_pr/pyproject.toml
[project]
name = "heart-disease-pr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "feature-engine>=1.9.3",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "scikit-learn==1.6.1",
    "streamlit>=1.49.1",
]
